Optimize Engine.set_target_variable().
This removes the extra call to do_set_target_variable() because
it's possible for target variables to be set during the build phase.
The extra call added a decent amount of overhead. Addtionally,
rather than setting the target variables for each target (via the
for loop), the bjam interface "set-target-variable" rule allows
for taking in a list of targets. Thus, even more function calls are
avoided.